fix: Migrate `api-interface.js` to Typescript, fix incorrect @nextcloud/axios usage.
Fix wrong import order in `HowToTool`
